pyproject.toml
[tool.poetry]
name = "kaggle-walmart"
version = "0.1.0"
description = ""
authors = ["shimacos37 <naoto.shimakoshi@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.7.7,<3.10"
numpy = "1.19.5"
pandas = "1.2.3"
scikit-learn = "^0.24.1"
matplotlib = "^3.3.3"
lightgbm = "^3.1.1"
tqdm = "^4.56.0"
hydra-core = "^1.0.5"
google-cloud-storage = "^1.35.0"
pandas-gbq = "^0.14.1"
google-api-python-client = "^2.0.2"
oauth2client = "^4.1.3"
seaborn = "^0.11.1"
japanize-matplotlib = "^1.1.3"
matplotlib-venn = "^0.11.6"
darts = "^0.7.0"

[tool.poetry.dev-dependencies]
invoke = "^1.4.1"
black = "^19.10b0"
pytest = "^5.4.2"
flake8 = "^3.8.1"
pre-commit = "^2.4.0"
mypy = "^0.782"
memory-profiler = "^0.58.0"
ipykernel = "^5.5.5"
jupyterlab = "^3.0.15"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
